"Moved test function to the correct location."
